requirements.txt
einops==0.8.0
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
patool==1.12
reformer-pytorch==1.4.4
scikit-learn==1.2.2
scipy==1.10.1
sktime==0.16.1
sympy==1.11.1
# torch==2.2.2
tqdm==4.64.1
transformers==4.46.3
# torchvision==0.17.2
timm==1.0.15
umap-learn==0.5.7
seaborn==0.13.2
PyWavelets==1.4.1